requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-telegram-bot==21.10
requests==2.32.3
gunicorn==23.0.0
SQLAlchemy==2.0.30